refactor: Split MainActivity.kt into separate screen files
- Extracted 15+ Composable screens from MainActivity.kt (2000+ lines) into 9 organized files
- Reduced MainActivity.kt from ~2000 lines to 284 lines (86% reduction)
- Created logical file structure:
  - DashboardScreen.kt - Dashboard components
  - HistoryScreens.kt - History list and detail screens
  - LibraryScreen.kt - Navigation library
  - ProgramManagementScreens.kt - Program management
  - TemplateManagementScreens.kt - Template management
  - WorkoutLoggerScreens.kt - Workout logging with timer
  - ExerciseManagementScreens.kt - Exercise CRUD operations
  - PersonalRecordsScreen.kt - PR display
  - SettingsScreen.kt - App settings
  - VolumeAnalysisScreen.kt - Volume analysis charts
- Fixed KSP version compatibility (2.1.21-1.0.29)
- Added missing padding import to MainActivity.kt
- Maintained exact same functionality and navigation behavior
