+++ app/src/main/java/com/example/myworkoutlog/MainActivity.kt
@@ -7,31 +7,14 @@ import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
 import androidx.activity.viewModels
-import androidx.compose.animation.AnimatedVisibility
-import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.foundation.text.KeyboardOptions
+import androidx.compose.foundation.layout.padding
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.*
 import androidx.compose.material.icons.automirrored.filled.*
-import androidx.compose.material.icons.outlined.*
 import androidx.compose.material3.*
 import androidx.compose.runtime.*
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.focus.onFocusChanged
-import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.text.font.FontStyle
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.input.KeyboardType
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
@@ -39,24 +22,6 @@ import androidx.navigation.compose.composable
 import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
 import com.example.myworkoutlog.ui.theme.MyWorkoutLogTheme
-import com.patrykandpatrick.vico.compose.axis.horizontal.rememberBottomAxis
-import com.patrykandpatrick.vico.compose.axis.vertical.rememberStartAxis
-import com.patrykandpatrick.vico.compose.chart.Chart
-import com.patrykandpatrick.vico.compose.chart.line.lineChart
-import com.patrykandpatrick.vico.core.entry.ChartEntryModelProducer
-import com.patrykandpatrick.vico.core.entry.entryOf
-import com.patrykandpatrick.vico.core.axis.AxisPosition
-import com.patrykandpatrick.vico.core.axis.formatter.AxisValueFormatter
-import com.patrykandpatrick.vico.compose.chart.column.columnChart
-import com.patrykandpatrick.vico.compose.chart.scroll.rememberChartScrollState
-import com.patrykandpatrick.vico.compose.component.shape.shader.verticalGradient
-import com.patrykandpatrick.vico.compose.component.lineComponent
-import androidx.compose.foundation.layout.ExperimentalLayoutApi
-import kotlin.math.roundToInt
-import java.text.SimpleDateFormat
-import java.util.*
-import java.util.concurrent.TimeUnit
-
 
 class MainActivity : ComponentActivity() {
     private val exerciseViewModel: ExerciseViewModel by viewModels {
@@ -86,7 +51,7 @@ class MainActivity : ComponentActivity() {
     private val programViewModel: ProgramViewModel by viewModels {
         ProgramViewModelFactory(
             (application as WorkoutApplication).database.programTemplateDao(),
-            (application as WorkoutApplication).database.workoutTemplateDao() // Pass the template DAO
+            (application as WorkoutApplication).database.workoutTemplateDao()
         )
     }
 
@@ -195,7 +160,6 @@ fun AppNavHost(
         composable(Screen.History.route) {
             HistoryScreen(
                 viewModel = historyViewModel,
-                // Pass a navigation action to the history screen
                 onNavigateToWorkout = { workoutId ->
                     navController.navigate(Screen.HistoryDetail.createRoute(workoutId))
                 }
@@ -254,7 +218,7 @@ fun AppNavHost(
         composable(Screen.ManagePrograms.route) {
             ManageProgramsScreen(
                 programViewModel = programViewModel,
-                activeCycleViewModel = activeCycleViewModel, // Pass it here
+                activeCycleViewModel = activeCycleViewModel,
                 onNavigateToProgram = { programId ->
                     navController.navigate(Screen.ProgramEditor.createRoute(programId))
                 }
@@ -265,7 +229,6 @@ fun AppNavHost(
             ProgramEditorScreen(
                 programId = programId,
                 programViewModel = programViewModel,
-                // Pass the template view model as well
                 templateViewModel = templateViewModel,
                 onNavigateUp = { navController.navigateUp() }
             )
@@ -273,7 +236,6 @@ fun AppNavHost(
         composable(Screen.PersonalRecords.route) {
             PersonalRecordsScreen(
                 viewModel = prViewModel,
-                // Pass a navigation action to the PR screen
                 onNavigateToWorkout = { workoutId ->
                     navController.navigate(Screen.HistoryDetail.createRoute(workoutId))
                 }
@@ -287,8 +249,6 @@ fun AppNavHost(
         }
     }
 }
-
-
 
 @Composable
 fun AppBottomNavigationBar(navController: NavHostController) {
@@ -304,14 +264,11 @@ fun AppBottomNavigationBar(navController: NavHostController) {
         items.forEach { screen ->
             NavigationBarItem(
                 icon = {
-                    // This 'when' block is now exhaustive
                     when (screen) {
                         Screen.Dashboard -> Icon(Icons.Filled.Home, contentDescription = "Dashboard")
                         Screen.History -> Icon(Icons.Filled.DateRange, contentDescription = "History")
                         Screen.Library -> Icon(Icons.AutoMirrored.Filled.List, contentDescription = "Library")
-                        // The 'else' branch handles any other cases, satisfying the compiler.
-                        // We need this even though our 'items' list is specific.
-                        else -> {} // Do nothing for other screens not in the nav bar
+                        else -> {}
                     }
                 },
                 label = { Text(screen.route.replaceFirstChar { it.uppercase() }) },
@@ -324,1988 +281,6 @@ fun AppBottomNavigationBar(navController: NavHostController) {
                     }
                 }
             )
-        }
-    }
-}
-
-
-// --- Our Screens ---
-
-@Composable
-fun DashboardScreen(
-    historyViewModel: HistoryViewModel,
-    activeCycleViewModel: ActiveCycleViewModel,
-    programViewModel: ProgramViewModel,
-    navController: NavHostController
-) {
-    val activeCycle by activeCycleViewModel.activeCycle.collectAsStateWithLifecycle()
-
-    // Conditionally show the correct dashboard
-    if (activeCycle == null) {
-        NoActiveCycleDashboard(
-            historyViewModel = historyViewModel,
-            onNavigate = { route -> navController.navigate(route) }
-        )
-    } else {
-        ActiveCycleDashboard(
-            activeCycle = activeCycle!!,
-            programViewModel = programViewModel,
-            activeCycleViewModel = activeCycleViewModel,
-            navController = navController
-        )
-    }
-}
-
-// NEW: The dashboard for when a cycle is active
-@Composable
-fun ActiveCycleDashboard(
-    activeCycle: ActiveProgramCycle,
-    programViewModel: ProgramViewModel,
-    activeCycleViewModel: ActiveCycleViewModel,
-    navController: NavHostController
-) {
-    // Get the details of the program blueprint for our active cycle
-    val program by programViewModel.getProgramById(activeCycle.programTemplateId).collectAsState(initial = null)
-
-    LazyColumn(
-        modifier = Modifier
-            .fillMaxSize()
-            .padding(16.dp),
-        verticalArrangement = Arrangement.spacedBy(16.dp)
-    ) {
-        item {
-            Text(activeCycle.userCycleName, fontSize = 24.sp, fontWeight = FontWeight.Bold)
-            Text(activeCycle.programTemplateName, style = MaterialTheme.typography.titleMedium)
-            Spacer(modifier = Modifier.height(8.dp))
-            Button(onClick = { activeCycleViewModel.endCycle() }, modifier = Modifier.fillMaxWidth()) {
-                Text("End Current Cycle")
-            }
-        }
-
-        program?.weeks?.sortedBy { it.order }?.forEach { week ->
-            item {
-                Text(week.weekLabel, style = MaterialTheme.typography.titleLarge)
-                HorizontalDivider(modifier = Modifier.padding(vertical = 4.dp))
-            }
-            items(week.sessions.sortedBy { it.order }) { session ->
-                val sessionKey = "${week.id}_${session.id}"
-                val isCompleted = activeCycle.completedSessions.containsKey(sessionKey)
-                val workoutId = activeCycle.completedSessions[sessionKey]
-
-                Card(elevation = CardDefaults.cardElevation(2.dp)) {
-                    Row(
-                        modifier = Modifier.fillMaxWidth().padding(horizontal = 16.dp, vertical = 8.dp),
-                        verticalAlignment = Alignment.CenterVertically
-                    ) {
-                        Text(session.sessionName, modifier = Modifier.weight(1f))
-                        if (isCompleted && workoutId != null) {
-                            Button(
-                                onClick = {
-                                    navController.navigate(Screen.HistoryDetail.createRoute(workoutId))
-                                },
-                                colors = ButtonDefaults.buttonColors(
-                                    containerColor = MaterialTheme.colorScheme.secondary
-                                )
-                            ) { Text("View") }
-                        } else {
-                            Button(onClick = {
-                                val route = Screen.WorkoutLogger.createRoute(
-                                    templateId = session.workoutTemplateId,
-                                    cycleId = activeCycle.id.toString(),
-                                    weekId = week.id,
-                                    sessionId = session.id
-                                )
-                                navController.navigate(route)
-                            }) {
-                                Text("Start")
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-@Composable
-fun NoActiveCycleDashboard(
-    historyViewModel: HistoryViewModel,
-    onNavigate: (String) -> Unit
-) {
-    val loggedWorkouts by historyViewModel.allLoggedWorkouts.collectAsStateWithLifecycle()
-    val latestWorkout = loggedWorkouts.firstOrNull()
-
-    // --- NEW: Prepare data and labels for the chart ---
-    val chartData = loggedWorkouts
-        .filter { it.bodyweight != null && it.bodyweight > 0 }
-        .sortedBy { it.date }
-
-    val bodyweightEntries = chartData.mapIndexed { index, workout ->
-        entryOf(index.toFloat(), workout.bodyweight!!.toFloat())
-    }
-
-    val chartModelProducer = ChartEntryModelProducer(bodyweightEntries)
-
-    // --- NEW: Create a custom formatter for the bottom axis ---
-    val bottomAxisValueFormatter = AxisValueFormatter<AxisPosition.Horizontal.Bottom> { value, _ ->
-        // We use the integer part of the value as an index to get the date
-        val index = value.toInt()
-        if (index in chartData.indices) {
-            // Format the date string like "06/08" from "2025-06-08"
-            val date = chartData[index].date
-            date.substring(5).replace('-', '/')
-        } else {
-            ""
-        }
-    }
-
-    // --- NEW: Dynamic Greeting Logic ---
-    val calendar = Calendar.getInstance()
-    val greeting = when (calendar.get(Calendar.HOUR_OF_DAY)) {
-        in 0..11 -> "Good morning!"
-        in 12..16 -> "Good afternoon!"
-        else -> "Good evening!"
-    }
-
-    LazyColumn(
-        modifier = Modifier
-            .fillMaxSize()
-            .padding(16.dp),
-        verticalArrangement = Arrangement.spacedBy(16.dp)
-    ) {
-        item {
-            Text("Dashboard", fontSize = 24.sp, fontWeight = FontWeight.Bold)
-        }
-
-        item {
-            Card(
-                modifier = Modifier.fillMaxWidth(),
-                elevation = CardDefaults.cardElevation(2.dp)
-            ) {
-                Column(modifier = Modifier.padding(16.dp)) {
-                    // Use the new dynamic greeting
-                    Text(greeting, style = MaterialTheme.typography.titleMedium)
-                    Spacer(modifier = Modifier.height(8.dp))
-                    Text("Ready for your next session?", style = MaterialTheme.typography.bodyMedium)
-                }
-            }
-        }
-
-        // The chart now has its axes configured with the new formatter
-        if (bodyweightEntries.isNotEmpty()) {
-            item {
-                Text("Bodyweight Trend", style = MaterialTheme.typography.titleMedium)
-                Spacer(modifier = Modifier.height(8.dp))
-                Card(elevation = CardDefaults.cardElevation(2.dp)) {
-                    Chart(
-                        chart = lineChart(),
-                        chartModelProducer = chartModelProducer,
-                        startAxis = rememberStartAxis(
-                            title = "Bodyweight" // Add a title to the Y-axis
-                        ),
-                        bottomAxis = rememberBottomAxis(
-                            valueFormatter = bottomAxisValueFormatter, // Use our custom date formatter
-                            guideline = null // Hide vertical guidelines for a cleaner look
-                        ),
-                        modifier = Modifier.padding(16.dp)
-                    )
-                }
-            }
-        }
-
-
-        item {
-            Text("Start a new workout:", style = MaterialTheme.typography.titleMedium)
-            Spacer(modifier = Modifier.height(8.dp))
-            Button(
-                onClick = { onNavigate(Screen.Library.route) },
-                modifier = Modifier.fillMaxWidth()
-            ) {
-                Text("Choose from Template")
-            }
-        }
-
-        if (latestWorkout != null) {
-            item {
-                Text("Last Workout", style = MaterialTheme.typography.titleMedium)
-                Spacer(modifier = Modifier.height(8.dp))
-                Card(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .clickable {
-                            // Navigate to the detail screen for this specific workout
-                            onNavigate(Screen.HistoryDetail.createRoute(latestWorkout.id))
-                        },
-                    elevation = CardDefaults.cardElevation(2.dp)
-                ) {
-                    Column(modifier = Modifier.padding(16.dp)) {
-                        Text(latestWorkout.name ?: "Workout", fontWeight = FontWeight.Bold)
-                        Text(latestWorkout.date, style = MaterialTheme.typography.bodySmall)
-                        Spacer(modifier = Modifier.height(8.dp))
-                        Text("${latestWorkout.loggedExercises.size} exercises performed.", style = MaterialTheme.typography.bodyMedium)
-                    }
-                }
-            }
-        }
-    }
-}
-
-
-
-private fun calculateDurationMinutes(start: Long?, end: Long?): Long? {
-    if (start == null || end == null) return null
-    val diff = end - start
-    return TimeUnit.MILLISECONDS.toMinutes(diff)
-}
-
-@Composable
-fun HistoryScreen(
-    viewModel: HistoryViewModel,
-    onNavigateToWorkout: (String) -> Unit
-) {
-    val loggedWorkouts by viewModel.allLoggedWorkouts.collectAsStateWithLifecycle()
-
-    Column(modifier = Modifier.padding(16.dp)) {
-        Text("Workout History", fontSize = 24.sp, fontWeight = FontWeight.Bold)
-        Spacer(modifier = Modifier.height(16.dp))
-
-        if (loggedWorkouts.isEmpty()) {
-            Box(modifier = Modifier.fillMaxSize(), contentAlignment = Alignment.Center) {
-                Text("No workouts logged yet.")
-            }
-        } else {
-            LazyColumn(verticalArrangement = Arrangement.spacedBy(8.dp)) {
-                items(loggedWorkouts) { workout ->
-                    Card(
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            // In the future, this will navigate to a detail view
-                            .clickable { onNavigateToWorkout(workout.id) },
-                        elevation = CardDefaults.cardElevation(2.dp)
-                    ) {
-                        Column(Modifier.padding(16.dp)) {
-                            Text(workout.name ?: "Workout", fontWeight = FontWeight.Bold)
-                            Text(workout.date, style = MaterialTheme.typography.bodySmall)
-                            Spacer(modifier = Modifier.height(8.dp))
-                            // NEW: Show time details directly in the list
-                            Row(
-                                modifier = Modifier.fillMaxWidth(),
-                                horizontalArrangement = Arrangement.SpaceBetween
-                            ) {
-                                val duration = calculateDurationMinutes(workout.startTimestamp, workout.endTimestamp)
-                                Text(
-                                    text = "Duration: ${duration?.let { "$it min" } ?: "--"}",
-                                    style = MaterialTheme.typography.bodyMedium
-                                )
-                                Text(
-                                    text = workout.startTimestamp?.let { formatTimestampToTime(it) } ?: "",
-                                    style = MaterialTheme.typography.bodyMedium
-                                )
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-
-private fun formatTimestampToTime(timestamp: Long): String {
-    return SimpleDateFormat("h:mm a", Locale.getDefault()).format(Date(timestamp))
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun HistoryDetailScreen(
-    workoutId: String,
-    viewModel: HistoryViewModel,
-    onNavigateUp: () -> Unit
-) {
-    val workout by viewModel.getLoggedWorkoutById(workoutId).collectAsState(initial = null)
-
-    Scaffold(
-        topBar = {
-            TopAppBar(
-                title = { Text(workout?.name ?: "Workout Details") },
-                navigationIcon = {
-                    IconButton(onClick = onNavigateUp) {
-                        Icon(Icons.AutoMirrored.Filled.ArrowBack, contentDescription = "Back")
-                    }
-                }
-            )
-        }
-    ) { paddingValues ->
-        val currentWorkout = workout
-        if (currentWorkout == null) {
-            Box(modifier = Modifier.fillMaxSize(), contentAlignment = Alignment.Center) {
-                CircularProgressIndicator()
-            }
-        } else {
-            LazyColumn(
-                modifier = Modifier.padding(paddingValues),
-                contentPadding = PaddingValues(16.dp),
-                verticalArrangement = Arrangement.spacedBy(16.dp)
-            ) {
-                item {
-                    // Display the date prominently here
-                    Text(currentWorkout.date, style = MaterialTheme.typography.headlineSmall)
-                    Spacer(modifier = Modifier.height(16.dp))
-
-                    Row(
-                        modifier = Modifier.fillMaxWidth(),
-                        horizontalArrangement = Arrangement.SpaceAround
-                    ) {
-                        // Start Time
-                        Column(horizontalAlignment = Alignment.CenterHorizontally) {
-                            Text("Start Time", style = MaterialTheme.typography.labelMedium)
-                            Text(
-                                text = currentWorkout.startTimestamp?.let { formatTimestampToTime(it) } ?: "--:--",
-                                style = MaterialTheme.typography.titleLarge,
-                                fontWeight = FontWeight.Bold
-                            )
-                        }
-                        // End Time
-                        Column(horizontalAlignment = Alignment.CenterHorizontally) {
-                            Text("End Time", style = MaterialTheme.typography.labelMedium)
-                            Text(
-                                text = currentWorkout.endTimestamp?.let { formatTimestampToTime(it) } ?: "--:--",
-                                style = MaterialTheme.typography.titleLarge,
-                                fontWeight = FontWeight.Bold
-                            )
-                        }
-                        // Duration
-                        Column(horizontalAlignment = Alignment.CenterHorizontally) {
-                            Text("Duration", style = MaterialTheme.typography.labelMedium)
-                            val durationMinutes = if (currentWorkout.startTimestamp != null && currentWorkout.endTimestamp != null) {
-                                val diff = currentWorkout.endTimestamp - currentWorkout.startTimestamp
-                                TimeUnit.MILLISECONDS.toMinutes(diff)
-                            } else null
-                            Text(
-                                text = durationMinutes?.let { "$it min" } ?: "- min",
-                                style = MaterialTheme.typography.titleLarge,
-                                fontWeight = FontWeight.Bold
-                            )
-                        }
-                    }
-                    HorizontalDivider(modifier = Modifier.padding(top = 16.dp))
-                }
-
-                // --- Overall Comments Section ---
-                if (!currentWorkout.overallComments.isNullOrBlank()) {
-                    item {
-                        Text("Notes:", style = MaterialTheme.typography.titleMedium)
-                        Text(currentWorkout.overallComments, style = MaterialTheme.typography.bodyMedium)
-                    }
-                }
-
-                items(currentWorkout.loggedExercises) { exercise ->
-                    Card(modifier = Modifier.fillMaxWidth()) {
-                        Column(modifier = Modifier.padding(16.dp)) {
-                            Text(exercise.exerciseName, fontSize = 18.sp, fontWeight = FontWeight.Bold)
-                            Spacer(Modifier.height(8.dp))
-
-                            // Display the header for the sets table
-                            Row(modifier = Modifier.fillMaxWidth()) {
-                                Text("Set", modifier = Modifier.width(60.dp), style = MaterialTheme.typography.labelSmall)
-                                Text("Weight (${workout!!.performedWeightUnit ?: "kg"})", modifier = Modifier.weight(1f), style = MaterialTheme.typography.labelSmall)
-                                Text("Reps", modifier = Modifier.weight(1f), style = MaterialTheme.typography.labelSmall)
-                            }
-                            HorizontalDivider(modifier = Modifier.padding(vertical = 4.dp))
-
-                            // Display each logged set
-                            exercise.sets.forEachIndexed { index, set ->
-                                Row(modifier = Modifier.fillMaxWidth()) {
-                                    Text((index + 1).toString(), modifier = Modifier.width(60.dp))
-                                    Text(set.weight?.toString() ?: "--", modifier = Modifier.weight(1f))
-                                    Text(set.reps?.toString() ?: "--", modifier = Modifier.weight(1f))
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-@Composable
-fun LibraryScreen(onNavigate: (String) -> Unit) {
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .padding(16.dp),
-        verticalArrangement = Arrangement.spacedBy(16.dp)
-    ) {
-        Text("Library", fontSize = 24.sp, fontWeight = FontWeight.Bold)
-        Card(
-            modifier = Modifier
-                .fillMaxWidth()
-                .clickable { onNavigate(Screen.ManageExercises.route) },
-            elevation = CardDefaults.cardElevation(2.dp)
-        ) {
-            Text("Manage Exercises", modifier = Modifier.padding(16.dp), fontSize = 18.sp)
-        }
-        Card(
-            modifier = Modifier
-                .fillMaxWidth()
-                .clickable { onNavigate(Screen.ManageTemplates.route) },
-            elevation = CardDefaults.cardElevation(2.dp)
-        ) {
-            Text("Manage Templates", modifier = Modifier.padding(16.dp), fontSize = 18.sp)
-        }
-        Card(
-            modifier = Modifier
-                .fillMaxWidth()
-                .clickable { onNavigate(Screen.ManagePrograms.route) },
-            elevation = CardDefaults.cardElevation(2.dp)
-        ) {
-            Text("Manage Program Blueprints", modifier = Modifier.padding(16.dp), fontSize = 18.sp)
-        }
-        Card(
-            modifier = Modifier
-                .fillMaxWidth()
-                .clickable { onNavigate(Screen.PersonalRecords.route) },
-            elevation = CardDefaults.cardElevation(2.dp)
-        ) {
-            Text("Personal Records", modifier = Modifier.padding(16.dp), fontSize = 18.sp)
-        }
-        Card(
-            modifier = Modifier
-                .fillMaxWidth()
-                .clickable { onNavigate(Screen.VolumeAnalysis.route) },
-            elevation = CardDefaults.cardElevation(2.dp)
-        ) {
-            Text("Volume Analysis", modifier = Modifier.padding(16.dp), fontSize = 18.sp)
-        }
-        Card(
-            modifier = Modifier
-                .fillMaxWidth()
-                .clickable { onNavigate(Screen.Settings.route) },
-            elevation = CardDefaults.cardElevation(2.dp)
-        ) {
-            Text("Settings", modifier = Modifier.padding(16.dp), fontSize = 18.sp)
-        }
-    }
-}
-
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun ManageProgramsScreen(
-    programViewModel: ProgramViewModel,
-    activeCycleViewModel: ActiveCycleViewModel,
-    onNavigateToProgram: (String) -> Unit
-) {
-    val programs by programViewModel.allPrograms.collectAsStateWithLifecycle()
-    var showCreateProgramDialog by remember { mutableStateOf(false) }
-    var showStartCycleDialog by remember { mutableStateOf<ProgramTemplate?>(null) }
-    var newName by remember { mutableStateOf("") }
-
-    Scaffold(
-        floatingActionButton = {
-            FloatingActionButton(onClick = { showCreateProgramDialog = true }) {
-                Icon(Icons.Filled.Add, contentDescription = "Create new program")
-            }
-        }
-    ) { paddingValues ->
-        Column(modifier = Modifier.padding(paddingValues).padding(16.dp)) {
-            Text("Program Blueprints", fontSize = 24.sp, fontWeight = FontWeight.Bold)
-            Spacer(modifier = Modifier.height(16.dp))
-
-            if (programs.isEmpty()) {
-                Text(
-                    "No programs yet. Click the '+' button to create one.",
-                    modifier = Modifier.align(Alignment.CenterHorizontally),
-                    textAlign = TextAlign.Center
-                )
-            } else {
-                LazyColumn(verticalArrangement = Arrangement.spacedBy(8.dp)) {
-                    items(programs) { program ->
-                        Card(elevation = CardDefaults.cardElevation(2.dp)) {
-                            Row(
-                                modifier = Modifier.padding(start = 16.dp),
-                                verticalAlignment = Alignment.CenterVertically
-                            ) {
-                                Column(
-                                    modifier = Modifier
-                                        .weight(1f)
-                                        .clickable { onNavigateToProgram(program.id) }
-                                        .padding(vertical = 16.dp)
-                                ) { Text(program.name) }
-                                // New "Start Cycle" button
-                                IconButton(onClick = { showStartCycleDialog = program }) {
-                                    Icon(
-                                        Icons.Outlined.PlayArrow,
-                                        contentDescription = "Start Cycle"
-                                    )
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-
-            // THIS IS THE DIALOG FOR CREATING A NEW PROGRAM
-            if (showCreateProgramDialog) {
-                AlertDialog(
-                    onDismissRequest = { showCreateProgramDialog = false },
-                    title = { Text("New Program Blueprint") },
-                    text = {
-                        OutlinedTextField(
-                            value = newName,
-                            onValueChange = { newName = it },
-                            label = { Text("Program Name") },
-                            singleLine = true
-                        )
-                    },
-                    confirmButton = {
-                        Button(
-                            onClick = {
-                                if (newName.isNotBlank()) {
-                                    programViewModel.insert(newName, null)
-                                    newName = ""
-                                    showCreateProgramDialog = false
-                                }
-                            }
-                        ) { Text("Create") }
-                    },
-                    dismissButton = {
-                        TextButton(onClick = { showCreateProgramDialog = false }) {
-                            Text("Cancel")
-                        }
-                    }
-                )
-            }
-
-            // THIS IS THE DIALOG FOR STARTING A NEW CYCLE
-            if (showStartCycleDialog != null) {
-                val programToStart = showStartCycleDialog!!
-                AlertDialog(
-                    onDismissRequest = { showStartCycleDialog = null },
-                    title = { Text("Start New Cycle") },
-                    text = {
-                        Column {
-                            Text("You are about to start the program: ${programToStart.name}")
-                            Spacer(Modifier.height(8.dp))
-                            OutlinedTextField(
-                                value = newName,
-                                onValueChange = { newName = it },
-                                label = { Text("Give this cycle a name") },
-                                placeholder = { Text(programToStart.name) }
-                            )
-                        }
-                    },
-                    confirmButton = {
-                        Button(onClick = {
-                            val cycleName = newName.ifBlank { programToStart.name }
-                            activeCycleViewModel.startCycle(programToStart, cycleName)
-                            showStartCycleDialog = null
-                            newName = ""
-                        }) { Text("Start") }
-                    },
-                    dismissButton = {
-                        TextButton(onClick = {
-                            showStartCycleDialog = null
-                        }) { Text("Cancel") }
-                    }
-                )
-            }
-        }
-    }
-}
-
-
-// NEW SCREEN for editing a program
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun ProgramEditorScreen(
-    programId: String,
-    programViewModel: ProgramViewModel,
-    templateViewModel: WorkoutTemplateViewModel, // New parameter
-    onNavigateUp: () -> Unit
-) {
-    val programFromDb by programViewModel.getProgramById(programId).collectAsState(initial = null)
-    val allWorkoutTemplates by templateViewModel.allTemplates.collectAsStateWithLifecycle()
-
-    var editedName by remember { mutableStateOf("") }
-    var editedWeeks by remember { mutableStateOf<List<ProgramWeekDefinition>>(emptyList()) }
-    var showAddSessionDialog by remember { mutableStateOf<String?>(null) } // Holds the ID of the week we're adding a session to
-
-    LaunchedEffect(programFromDb) {
-        programFromDb?.let {
-            editedName = it.name
-            editedWeeks = it.weeks
-        }
-    }
-
-    Scaffold(
-        topBar = {
-            TopAppBar(
-                title = { Text(editedName) },
-                navigationIcon = {
-                    IconButton(onClick = onNavigateUp) {
-                        Icon(Icons.AutoMirrored.Filled.ArrowBack, contentDescription = "Back")
-                    }
-                },
-                actions = {
-                    Button(onClick = {
-                        programFromDb?.let {
-                            val updatedProgram = it.copy(name = editedName, weeks = editedWeeks)
-                            programViewModel.update(updatedProgram)
-                        }
-                        onNavigateUp()
-                    }) { Text("Save") }
-                }
-            )
-        },
-        floatingActionButton = {
-            FloatingActionButton(onClick = {
-                val newWeek = ProgramWeekDefinition(
-                    id = UUID.randomUUID().toString(),
-                    weekLabel = "Week ${editedWeeks.size + 1}",
-                    order = editedWeeks.size + 1,
-                    sessions = emptyList()
-                )
-                editedWeeks = editedWeeks + newWeek
-            }) {
-                Icon(Icons.Filled.Add, "Add Week")
-            }
-        }
-    ) { paddingValues ->
-        Column(modifier = Modifier.padding(paddingValues).padding(horizontal = 16.dp)) {
-            OutlinedTextField(
-                value = editedName,
-                onValueChange = { editedName = it },
-                label = { Text("Program Name") },
-                modifier = Modifier.fillMaxWidth().padding(vertical = 8.dp)
-            )
-            LazyColumn(verticalArrangement = Arrangement.spacedBy(16.dp)) {
-                items(editedWeeks) { week ->
-                    Card(elevation = CardDefaults.cardElevation(2.dp)) {
-                        Column(Modifier.padding(16.dp)) {
-                            Row(verticalAlignment = Alignment.CenterVertically) {
-                                OutlinedTextField(
-                                    value = week.weekLabel,
-                                    onValueChange = { newLabel ->
-                                        editedWeeks = editedWeeks.map {
-                                            if (it.id == week.id) it.copy(weekLabel = newLabel) else it
-                                        }
-                                    },
-                                    label = { Text("Week Label") },
-                                    modifier = Modifier.weight(1f)
-                                )
-                                IconButton(onClick = {
-                                    editedWeeks = editedWeeks.filter { it.id != week.id }
-                                }) {
-                                    Icon(Icons.Default.Delete, contentDescription = "Delete Week")
-                                }
-                            }
-                            Spacer(Modifier.height(8.dp))
-                            // List the sessions for this week
-                            week.sessions.sortedBy { it.order }.forEach { session ->
-                                val templateName = allWorkoutTemplates.find { it.id == session.workoutTemplateId }?.name ?: "Unknown Template"
-                                Text("  - ${session.sessionName}: $templateName", style = MaterialTheme.typography.bodyMedium)
-                            }
-                            TextButton(onClick = { showAddSessionDialog = week.id }) {
-                                Text("Add Session to Week")
-                            }
-                        }
-                    }
-                }
-            }
-        }
-
-        // Dialog for adding a session to a week
-        if (showAddSessionDialog != null) {
-            val weekIdToAddSessionTo = showAddSessionDialog
-            var sessionName by remember { mutableStateOf("") }
-            AlertDialog(
-                onDismissRequest = { showAddSessionDialog = null },
-                title = { Text("Add Session") },
-                text = {
-                    Column {
-                        OutlinedTextField(
-                            value = sessionName,
-                            onValueChange = { sessionName = it },
-                            label = { Text("Session Name (e.g., Day 1)") }
-                        )
-                        Spacer(Modifier.height(8.dp))
-                        Text("Choose a template:", fontWeight = FontWeight.Bold)
-                        LazyColumn(modifier = Modifier.height(150.dp)) {
-                            items(allWorkoutTemplates) { template ->
-                                Text(
-                                    text = template.name,
-                                    modifier = Modifier
-                                        .fillMaxWidth()
-                                        .clickable {
-                                            val weekToUpdate = editedWeeks.find { it.id == weekIdToAddSessionTo }
-                                            if (weekToUpdate != null) {
-                                                val newSession = ProgramSessionDefinition(
-                                                    id = UUID.randomUUID().toString(),
-                                                    sessionName = sessionName.ifBlank { "Session ${weekToUpdate.sessions.size + 1}" },
-                                                    workoutTemplateId = template.id,
-                                                    order = weekToUpdate.sessions.size + 1
-                                                )
-                                                val updatedSessions = weekToUpdate.sessions + newSession
-                                                val updatedWeek = weekToUpdate.copy(sessions = updatedSessions)
-                                                editedWeeks = editedWeeks.map { if (it.id == weekIdToAddSessionTo) updatedWeek else it }
-                                            }
-                                            showAddSessionDialog = null
-                                        }
-                                        .padding(vertical = 8.dp)
-                                )
-                            }
-                        }
-                    }
-                },
-                confirmButton = {
-                    TextButton(onClick = { showAddSessionDialog = null }) { Text("Cancel") }
-                }
-            )
-        }
-    }
-}
-
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun ManageTemplatesScreen(
-    viewModel: WorkoutTemplateViewModel,
-    onNavigateToTemplate: (String) -> Unit,
-    onStartWorkout: (String) -> Unit
-) {
-    val templates by viewModel.allTemplates.collectAsStateWithLifecycle()
-    var showDialog by remember { mutableStateOf(false) }
-    var templateName by remember { mutableStateOf("") }
-
-    Scaffold(
-        floatingActionButton = {
-            FloatingActionButton(onClick = { showDialog = true }) {
-                Icon(Icons.Filled.Add, contentDescription = "Create new template")
-            }
-        }
-    ) { paddingValues ->
-        Column(modifier = Modifier.padding(paddingValues).padding(16.dp)) {
-            Text("Workout Templates", fontSize = 24.sp, fontWeight = FontWeight.Bold)
-            Spacer(modifier = Modifier.height(16.dp))
-
-            if (templates.isEmpty()) {
-                Text(
-                    "No templates yet. Click the '+' button to create one.",
-                    modifier = Modifier.align(Alignment.CenterHorizontally)
-                )
-            } else {
-                LazyColumn(verticalArrangement = Arrangement.spacedBy(8.dp)) {
-                    items(templates) { template ->
-                        Card(
-                            modifier = Modifier.fillMaxWidth(),
-                            elevation = CardDefaults.cardElevation(2.dp)
-                        ) {
-                            Row(
-                                modifier = Modifier.padding(start = 16.dp),
-                                verticalAlignment = Alignment.CenterVertically
-                            ) {
-                                Column(
-                                    modifier = Modifier
-                                        .weight(1f)
-                                        .clickable { onNavigateToTemplate(template.id) }
-                                        .padding(vertical = 16.dp)
-                                ) {
-                                    Text(template.name)
-                                }
-                                // ADDED "START" BUTTON
-                                IconButton(
-                                    onClick = { onStartWorkout(template.id) },
-                                    modifier = Modifier.padding(16.dp)
-                                ) {
-                                    Icon(
-                                        Icons.Filled.PlayArrow,
-                                        contentDescription = "Start Workout"
-                                    )
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-
-            if (showDialog) {
-                AlertDialog(
-                    onDismissRequest = { showDialog = false },
-                    title = { Text("New Workout Template") },
-                    text = {
-                        OutlinedTextField(
-                            value = templateName,
-                            onValueChange = { templateName = it },
-                            label = { Text("Template Name") },
-                            singleLine = true
-                        )
-                    },
-                    confirmButton = {
-                        Button(
-                            onClick = {
-                                if (templateName.isNotBlank()) {
-                                    viewModel.insert(templateName, null)
-                                    templateName = ""
-                                    showDialog = false
-                                }
-                            }
-                        ) { Text("Create") }
-                    },
-                    dismissButton = {
-                        TextButton(onClick = { showDialog = false }) {
-                            Text("Cancel")
-                        }
-                    }
-                )
-            }
-        }
-    }
-}
-
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun WorkoutLoggerScreen(
-    templateId: String,
-    cycleId: String?,
-    weekId: String?,
-    sessionId: String?,
-    viewModel: WorkoutLoggerViewModel,
-    activeCycleViewModel: ActiveCycleViewModel,
-    weightUnit: String,
-    onNavigateUp: () -> Unit
-) {
-    // Local state for the bodyweight text field
-    var bodyweightText by remember { mutableStateOf("") }
-
-    // LaunchedEffect runs a coroutine when the composable first appears.
-    // We use it to tell the ViewModel to load our template.
-    // The 'key1 = templateId' means it will only re-run if the templateId changes.
-    LaunchedEffect(key1 = templateId) {
-        viewModel.startWorkoutFromTemplate(templateId, cycleId, weekId, sessionId)
-    }
-
-    // Collect the active workout state from the ViewModel.
-    // The UI will automatically recompose whenever this state changes.
-    val activeWorkout by viewModel.activeWorkoutState.collectAsStateWithLifecycle()
-    val activeCycle by activeCycleViewModel.activeCycle.collectAsStateWithLifecycle()
-
-    // Get the timer state from the ViewModel
-    val timerIsRunning by viewModel.timerIsRunning.collectAsStateWithLifecycle()
-    val timerValue by viewModel.timerValueSeconds.collectAsStateWithLifecycle()
-    val sessionElapsedTime by viewModel.sessionElapsedTime.collectAsStateWithLifecycle()
-
-
-    Scaffold(
-        topBar = {
-            TopAppBar(
-                title = { Text(activeWorkout?.name ?: "Log Workout") },
-                navigationIcon = {
-                    IconButton(onClick = {
-                        viewModel.stopRestTimer()
-                        onNavigateUp()
-                    }) {
-                        Icon(Icons.AutoMirrored.Filled.ArrowBack, contentDescription = "Back")
-                    }
-                },
-                actions = {
-                    Text(
-                        text = formatTime(sessionElapsedTime),
-                        style = MaterialTheme.typography.titleMedium,
-                        modifier = Modifier.padding(end = 8.dp)
-                    )
-                    Button(onClick = {
-                        viewModel.finishWorkout(weightUnit, activeCycle)
-                        onNavigateUp()
-                    }) {
-                        Text("Finish")
-                    }
-                }
-            )
-        },
-
-        bottomBar = {
-            TimerBar(
-                isRunning = timerIsRunning,
-                currentTime = timerValue,
-                onPause = { viewModel.pauseRestTimer() },
-                onResume = { viewModel.resumeRestTimer() },
-                onStop = { viewModel.stopRestTimer() },
-                onReset = { viewModel.resetRestTimer() },
-                onAddTime = { viewModel.addTimeToRestTimer(15) }
-            )
-        }
-    ) { paddingValues ->
-        // If the workout is not loaded yet, show a loading indicator.
-        if (activeWorkout == null) {
-            Box(modifier = Modifier.fillMaxSize(), contentAlignment = Alignment.Center) {
-                CircularProgressIndicator()
-            }
-        } else {
-            // Once loaded, display the list of exercises.
-            LazyColumn(
-                modifier = Modifier
-                    .padding(paddingValues)
-                    .padding(16.dp),
-                verticalArrangement = Arrangement.spacedBy(16.dp)
-            ) {
-                // NEW: Add a bodyweight input field as the first item in the list
-                item {
-                    OutlinedTextField(
-                        value = bodyweightText,
-                        onValueChange = { newText ->
-                            if (newText.matches(Regex("^\\d*\\.?\\d*\$"))) {
-                                bodyweightText = newText
-                            }
-                        },
-                        label = { Text("Bodyweight ($weightUnit)") },
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .onFocusChanged { focusState ->
-                                if (!focusState.isFocused) {
-                                    viewModel.updateBodyweight(bodyweightText)
-                                }
-                            }
-                    )
-                }
-                items(activeWorkout!!.loggedExercises) { exercise ->
-                    Card(modifier = Modifier.fillMaxWidth()) {
-                        Column(modifier = Modifier.padding(16.dp)) {
-                            Text(exercise.exerciseName, fontSize = 18.sp, fontWeight = FontWeight.Bold)
-                            Spacer(Modifier.height(8.dp))
-                            HorizontalDivider(modifier = Modifier.padding(vertical = 8.dp))
-
-                            // For each exercise, display its sets
-                            exercise.sets.forEachIndexed { index, set ->
-                                LoggedSetRow(
-                                    set = set,
-                                    setNumber = index + 1,
-                                    weightUnit = weightUnit,
-                                    onRepsChange = { newReps ->
-                                        viewModel.updateSet(exercise.id, set.id, newReps, set.weight, set.secs?.toString() ?: "")
-                                    },
-                                    onWeightChange = { newWeight ->
-                                        viewModel.updateSet(exercise.id, set.id, set.reps?.toString() ?: "", newWeight, set.secs?.toString() ?: "")
-                                    },
-                                    onSecsChange = { newSecs ->
-                                        viewModel.updateSet(exercise.id, set.id, set.reps?.toString() ?: "", set.weight, newSecs)
-                                    },
-                                    // Pass the ViewModel function to the button
-                                    onStartRest = { viewModel.startRestTimer() }
-                                )
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun TemplateDetailScreen(
-    templateId: String,
-    viewModel: WorkoutTemplateViewModel,
-    onNavigateUp: () -> Unit
-) {
-    val templateFromDb by viewModel.getTemplateById(templateId).collectAsState(initial = null)
-    val allExercises by viewModel.allMasterExercises.collectAsStateWithLifecycle()
-
-    var editedName by remember { mutableStateOf("") }
-    var editedExercises by remember { mutableStateOf<List<TemplateExercise>>(emptyList()) }
-    var showAddExerciseDialog by remember { mutableStateOf(false) }
-
-    // This effect now runs only when the template from the database (templateFromDb)
-    // changes, and it updates the local, editable state.
-    LaunchedEffect(templateFromDb) {
-        templateFromDb?.let {
-            // This check prevents overwriting user's current edits when the DB updates
-            //if (it.templateExercises != editedExercises) {
-                editedName = it.name
-                editedExercises = it.templateExercises
-            //}
-        }
-    }
-
-    Scaffold(
-        topBar = {
-            TopAppBar(
-                title = { Text(editedName) },
-                navigationIcon = {
-                    IconButton(onClick = onNavigateUp) {
-                        Icon(Icons.AutoMirrored.Filled.ArrowBack, contentDescription = "Back")
-                    }
-                },
-                actions = {
-                    Button(onClick = {
-                        // When saving, use the current state of our editable variables
-                        templateFromDb?.let {
-                            val updatedTemplate = it.copy(
-                                name = editedName,
-                                templateExercises = editedExercises
-                            )
-                            viewModel.update(updatedTemplate)
-                        }
-                        onNavigateUp()
-                    }) {
-                        Text("Save")
-                    }
-                }
-            )
-        },
-        floatingActionButton = {
-            FloatingActionButton(onClick = { showAddExerciseDialog = true }) {
-                Icon(Icons.Filled.Add, "Add Exercise")
-            }
-        }
-    ) { paddingValues ->
-        Column(
-            modifier = Modifier
-                .padding(paddingValues)
-                .padding(16.dp),
-            verticalArrangement = Arrangement.spacedBy(16.dp)
-        ) {
-            OutlinedTextField(
-                value = editedName,
-                onValueChange = { editedName = it },
-                label = { Text("Template Name") },
-                modifier = Modifier.fillMaxWidth()
-            )
-
-            LazyColumn(verticalArrangement = Arrangement.spacedBy(16.dp)) {
-                items(editedExercises, key = { it.id }) { templateExercise ->
-                    Card(modifier = Modifier.fillMaxWidth(), elevation = CardDefaults.cardElevation(2.dp)) {
-                        Column(modifier = Modifier.padding(16.dp)) {
-                            Text(templateExercise.exerciseName, fontWeight = FontWeight.Bold, fontSize = 18.sp)
-                            HorizontalDivider(modifier = Modifier.padding(vertical = 8.dp))
-
-                            // Loop through each set and make its targets editable
-                            templateExercise.sets.forEachIndexed { setIndex, set ->
-                                TemplateSetEditorRow(
-                                    set = set,
-                                    setNumber = setIndex + 1,
-                                    onSetChange = { updatedSet ->
-                                        val newSets = templateExercise.sets.toMutableList()
-                                        newSets[setIndex] = updatedSet
-                                        val newExercise = templateExercise.copy(sets = newSets)
-                                        editedExercises = editedExercises.map { if (it.id == newExercise.id) newExercise else it }
-                                    },
-                                    onDelete = {
-                                        val newSets = templateExercise.sets.toMutableList()
-                                        newSets.removeAt(setIndex)
-                                        val newExercise = templateExercise.copy(sets = newSets)
-                                        editedExercises = editedExercises.map { if (it.id == newExercise.id) newExercise else it }
-                                    }
-                                )
-                            }
-
-                            TextButton(
-                                onClick = {
-                                    val newSet = TemplateExerciseSet(id = UUID.randomUUID().toString(), targetReps = null, targetSecs = null)
-                                    val newExercise = templateExercise.copy(sets = templateExercise.sets + newSet)
-                                    editedExercises = editedExercises.map { if (it.id == newExercise.id) newExercise else it }
-                                },
-                                modifier = Modifier.align(Alignment.End)
-                            ) { Text("Add Set") }
-                        }
-                    }
-                }
-            }
-        }
-
-        if (showAddExerciseDialog) {
-            AlertDialog(
-                onDismissRequest = { showAddExerciseDialog = false },
-                title = { Text("Add Exercise to Template") },
-                text = {
-                    LazyColumn {
-                        items(allExercises) { exercise ->
-                            Text(
-                                text = exercise.name,
-                                modifier = Modifier
-                                    .fillMaxWidth()
-                                    .clickable {
-                                        val newTemplateExercise = TemplateExercise(
-                                            id = UUID.randomUUID().toString(),
-                                            exerciseId = exercise.id,
-                                            exerciseName = exercise.name,
-                                            targetMuscleGroups = exercise.targetMuscleGroups,
-                                            equipment = exercise.equipment,
-                                            order = (editedExercises.maxOfOrNull { it.order } ?: 0) + 1,
-                                            sets = listOf(TemplateExerciseSet(id = UUID.randomUUID().toString()))
-                                        )
-                                        editedExercises = editedExercises + newTemplateExercise
-                                        showAddExerciseDialog = false
-                                    }
-                                    .padding(vertical = 8.dp)
-                            )
-                        }
-                    }
-                },
-                confirmButton = { TextButton(onClick = { showAddExerciseDialog = false }) { Text("Cancel") } }
-            )
-        }
-    }
-}
-
-
-// NEW HELPER COMPOSABLE for editing a set in a template
-@Composable
-fun TemplateSetEditorRow(
-    set: TemplateExerciseSet,
-    setNumber: Int,
-    onSetChange: (TemplateExerciseSet) -> Unit,
-    onDelete: () -> Unit
-) {
-    Row(
-        modifier = Modifier.fillMaxWidth().padding(vertical = 4.dp),
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.spacedBy(8.dp)
-    ) {
-        Text("Set $setNumber", modifier = Modifier.width(60.dp))
-        OutlinedTextField(
-            value = set.targetReps ?: "",
-            onValueChange = { onSetChange(set.copy(targetReps = it, targetSecs = null)) },
-            label = { Text("Reps") },
-            modifier = Modifier.weight(1f)
-        )
-        OutlinedTextField(
-            value = set.targetSecs ?: "",
-            onValueChange = { onSetChange(set.copy(targetSecs = it, targetReps = null)) },
-            label = { Text("Secs") },
-            modifier = Modifier.weight(1f)
-        )
-        IconButton(onClick = onDelete) {
-            Icon(Icons.Default.Delete, contentDescription = "Delete Set")
-        }
-    }
-}
-
-
-@Composable
-fun PersonalRecordsScreen(
-    viewModel: PrViewModel,
-    onNavigateToWorkout: (String) -> Unit
-) {
-    val searchText by viewModel.searchText.collectAsStateWithLifecycle()
-    val filteredPRs by viewModel.filteredPRs.collectAsStateWithLifecycle()
-    val prsByExercise = filteredPRs.groupBy { it.exerciseName }
-
-    Column(modifier = Modifier.fillMaxSize()) {
-        OutlinedTextField(
-            value = searchText,
-            onValueChange = viewModel::onSearchTextChanged,
-            label = { Text("Search Exercise...") },
-            modifier = Modifier.fillMaxWidth().padding(16.dp),
-            singleLine = true,
-            leadingIcon = { Icon(Icons.Filled.Search, contentDescription = "Search") }
-        )
-
-        if (prsByExercise.isEmpty()) {
-            Box(modifier = Modifier.fillMaxSize(), contentAlignment = Alignment.Center) {
-                Text("No PRs found.", style = MaterialTheme.typography.bodyLarge)
-            }
-        } else {
-            LazyColumn(
-                contentPadding = PaddingValues(horizontal = 16.dp, vertical = 8.dp),
-                verticalArrangement = Arrangement.spacedBy(16.dp)
-            ) {
-                prsByExercise.forEach { (exerciseName, prsForExercise) ->
-                    item {
-                        Card(
-                            modifier = Modifier.fillMaxWidth(),
-                            elevation = CardDefaults.cardElevation(2.dp)
-                        ) {
-                            Column(modifier = Modifier.padding(16.dp)) {
-                                Text(
-                                    exerciseName,
-                                    style = MaterialTheme.typography.titleLarge,
-                                    fontWeight = FontWeight.Bold
-                                )
-                                Spacer(modifier = Modifier.height(8.dp))
-
-                                // Separate PRs by type for clarity
-                                val weightPRs = prsForExercise.filter { it.type == PRType.MAX_WEIGHT_FOR_REPS }.sortedBy { it.reps }
-                                val repsPRs = prsForExercise.filter { it.type == PRType.MAX_REPS_AT_WEIGHT }.sortedByDescending { it.weight }
-                                val durationPRs = prsForExercise.filter { it.type == PRType.DURATION }
-
-                                // Display each PR type in its own row with an icon
-                                if (weightPRs.isNotEmpty()) {
-                                    HorizontalDivider(modifier = Modifier.padding(vertical = 8.dp))
-                                    PRCategoryRow(icon = Icons.Filled.FitnessCenter, title = "Best Weight for Reps")
-                                    weightPRs.forEach { pr ->
-                                        PRDetailRow(pr = pr, onNavigateToWorkout = onNavigateToWorkout)
-                                    }
-                                }
-                                if (repsPRs.isNotEmpty()) {
-                                    HorizontalDivider(modifier = Modifier.padding(vertical = 8.dp))
-                                    PRCategoryRow(icon = Icons.Filled.Repeat, title = "Best Reps at Weight")
-                                    repsPRs.forEach { pr ->
-                                        PRDetailRow(pr = pr, onNavigateToWorkout = onNavigateToWorkout)
-                                    }
-                                }
-                                if (durationPRs.isNotEmpty()) {
-                                    HorizontalDivider(modifier = Modifier.padding(vertical = 8.dp))
-                                    PRCategoryRow(icon = Icons.Filled.Timer, title = "Best Duration")
-                                    durationPRs.forEach { pr ->
-                                        PRDetailRow(pr = pr, onNavigateToWorkout = onNavigateToWorkout)
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-// NEW Helper Composable for Category Headers
-@Composable
-fun PRCategoryRow(icon: ImageVector, title: String) {
-    Row(verticalAlignment = Alignment.CenterVertically, horizontalArrangement = Arrangement.spacedBy(8.dp)) {
-        Icon(imageVector = icon, contentDescription = title, tint = MaterialTheme.colorScheme.primary)
-        Text(text = title, style = MaterialTheme.typography.titleMedium)
-    }
-}
-
-// NEW Helper Composable for PR Details
-@Composable
-fun PRDetailRow(pr: PersonalRecord, onNavigateToWorkout: (String) -> Unit) {
-    val weightUnit = pr.weightUnit ?: "kg"
-    val e1RM = if (pr.weight != null && pr.reps != null) {
-        StrengthAnalytics.calculateEpley1RM(pr.weight, pr.reps).roundToInt()
-    } else null
-
-    val prText = when (pr.type) {
-        PRType.MAX_WEIGHT_FOR_REPS -> "${pr.reps} reps"
-        PRType.MAX_REPS_AT_WEIGHT -> "@ ${pr.weight} $weightUnit"
-        PRType.DURATION -> "Max Duration"
-    }
-    val prValue = when (pr.type) {
-        PRType.MAX_WEIGHT_FOR_REPS -> "${pr.weight} $weightUnit"
-        PRType.MAX_REPS_AT_WEIGHT -> "${pr.reps} reps"
-        PRType.DURATION -> "${pr.durationSecs}s"
-    }
-
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .clickable { onNavigateToWorkout(pr.loggedWorkoutId) }
-            .padding(top = 8.dp, bottom = 4.dp, start = 8.dp), // Indent PRs
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        Column(modifier = Modifier.weight(1f)) {
-            Text(prText, style = MaterialTheme.typography.bodyLarge)
-            Text(pr.date, style = MaterialTheme.typography.bodySmall, fontStyle = FontStyle.Italic)
-        }
-        Column(horizontalAlignment = Alignment.End) {
-            Text(prValue, style = MaterialTheme.typography.titleMedium, fontWeight = FontWeight.Bold)
-            if (e1RM != null && pr.type == PRType.MAX_WEIGHT_FOR_REPS) {
-                Text("e1RM: $e1RM $weightUnit", style = MaterialTheme.typography.bodySmall, fontStyle = FontStyle.Italic)
-            }
-        }
-    }
-}
-
-
-@Composable
-fun SettingsScreen(viewModel: SettingsViewModel) {
-    val weightUnit by viewModel.weightUnit.collectAsStateWithLifecycle()
-    val options = listOf("kg", "lb")
-
-    Column(modifier = Modifier.padding(16.dp)) {
-        Text("Settings", fontSize = 24.sp, fontWeight = FontWeight.Bold)
-        Spacer(modifier = Modifier.height(16.dp))
-
-        Row(
-            modifier = Modifier.fillMaxWidth(),
-            verticalAlignment = Alignment.CenterVertically,
-            horizontalArrangement = Arrangement.SpaceBetween
-        ) {
-            Text("Weight Unit", style = MaterialTheme.typography.bodyLarge)
-            SingleChoiceSegmentedButtonRow {
-                options.forEachIndexed { index, label ->
-                    SegmentedButton(
-                        shape = SegmentedButtonDefaults.itemShape(index = index, count = options.size),
-                        onClick = { viewModel.setWeightUnit(label) },
-                        selected = weightUnit == label
-                    ) {
-                        Text(label)
-                    }
-                }
-            }
-        }
-    }
-}
-
-
-@Composable
-fun LoggedSetRow(
-    set: LoggedSet,
-    setNumber: Int,
-    weightUnit: String,
-    onRepsChange: (String) -> Unit,
-    onWeightChange: (Double?) -> Unit,
-    onSecsChange: (String) -> Unit,
-    onStartRest: () -> Unit
-) {
-    // Determine which fields to show based on the template's targets for this set
-    val showWeightReps = !set.targetReps.isNullOrBlank()
-    val showSecs = !set.targetSecs.isNullOrBlank()
-
-    // This local state holds the text as the user types it.
-    var weightText by remember { mutableStateOf(set.weight?.toString() ?: "") }
-    var repsText by remember { mutableStateOf(set.reps?.toString() ?: "") }
-    var secsText by remember { mutableStateOf(set.secs?.toString() ?: "") }
-
-    // This ensures if the underlying data changes from elsewhere, our text fields update.
-    LaunchedEffect(set.weight) {
-        val currentWeightString = set.weight?.toString() ?: ""
-        if (weightText != currentWeightString) {
-            weightText = currentWeightString
-        }
-    }
-    LaunchedEffect(set.reps) {
-        val currentRepsString = set.reps?.toString() ?: ""
-        if (repsText != currentRepsString) {
-            repsText = currentRepsString
-        }
-    }
-    LaunchedEffect(set.secs) {
-        val currentSecsString = set.secs?.toString() ?: ""
-        if (secsText != currentSecsString) {
-            secsText = currentSecsString
-        }
-    }
-
-    Column(modifier = Modifier.padding(vertical = 4.dp)) {
-        Row(
-            modifier = Modifier.fillMaxWidth(),
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            Text(
-                text = "Set $setNumber",
-                modifier = Modifier.width(60.dp),
-                fontWeight = FontWeight.Bold
-            )
-            Spacer(modifier = Modifier.weight(1f))
-            if (set.reps != null || set.secs != null) {
-                IconButton(onClick = onStartRest) {
-                    Icon(Icons.Filled.PlayArrow, contentDescription = "Start Rest Timer")
-                }
-            }
-        }
-
-        Row(
-            modifier = Modifier.fillMaxWidth(),
-            verticalAlignment = Alignment.CenterVertically,
-            horizontalArrangement = Arrangement.spacedBy(8.dp)
-        ) {
-            // Text field for weight
-            OutlinedTextField(
-                value = weightText,
-                // onValueChange now just updates the local text state
-                onValueChange = { newText ->
-                    // Allow only digits and one decimal point
-                    if (newText.matches(Regex("^\\d*\\.?\\d*\$"))) {
-                        weightText = newText
-                    }
-                },
-                label = { Text("Weight ($weightUnit)") },
-                keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Number),
-                modifier = Modifier
-                    .weight(1f)
-                    // This is the magic: when you tap away, it saves the value.
-                    .onFocusChanged { focusState ->
-                        if (!focusState.isFocused) {
-                            onWeightChange(weightText.toDoubleOrNull())
-                        }
-                    }
-            )
-
-            if (showWeightReps) {
-                // Text field for reps
-                OutlinedTextField(
-                    value = repsText,
-                    onValueChange = { newText ->
-                        if (newText.all { it.isDigit() }) {
-                            repsText = newText
-                        }
-                    },
-                    label = { Text("Reps") },
-                    keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Number),
-                    modifier = Modifier
-                        .weight(1f)
-                        .onFocusChanged { focusState ->
-                            if (!focusState.isFocused) {
-                                onRepsChange(repsText)
-                            }
-                        }
-                )
-            }
-
-            // This TextField will only appear for time-based exercises
-            if (showSecs) {
-                OutlinedTextField(
-                    value = secsText,
-                    onValueChange = { newText ->
-                        if (newText.all { it.isDigit() }) {
-                            secsText = newText
-                        }
-                    },
-                    label = { Text("Secs") },
-                    keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Number),
-                    modifier = Modifier
-                        .weight(1f)
-                        .onFocusChanged { focusState ->
-                            if (!focusState.isFocused) {
-                                onSecsChange(secsText)
-                            }
-                        }
-                )
-            }
-        }
-    }
-}
-
-
-
-@Composable
-fun ManageExercisesScreen(viewModel: ExerciseViewModel) {
-    val exercises by viewModel.allExercises.collectAsStateWithLifecycle()
-
-    // State for managing the dialogs
-    var showAddDialog by remember { mutableStateOf(false) }
-    var exerciseToEdit by remember { mutableStateOf<Exercise?>(null) }
-    var exerciseToDelete by remember { mutableStateOf<Exercise?>(null) }
-
-    // Scaffold provides the structure for the FloatingActionButton
-    Scaffold(
-        floatingActionButton = {
-            FloatingActionButton(onClick = { showAddDialog = true }) {
-                Icon(Icons.Filled.Add, contentDescription = "Add new exercise")
-            }
-        }
-    ) { paddingValues ->
-        Column(
-            modifier = Modifier
-                .padding(paddingValues)
-                .padding(16.dp)
-        ) {
-            Text("Manage Exercises", fontSize = 24.sp, fontWeight = FontWeight.Bold)
-            Spacer(modifier = Modifier.height(16.dp))
-
-            LazyColumn(verticalArrangement = Arrangement.spacedBy(8.dp)) {
-                items(exercises) { exercise ->
-                    ExerciseItem(
-                        exercise = exercise,
-                        onEditClick = { exerciseToEdit = exercise },
-                        onDeleteClick = { exerciseToDelete = exercise }
-                    )
-                }
-            }
-        }
-    }
-
-    // Show the Add dialog when showAddDialog is true
-    if (showAddDialog) {
-        AddExerciseDialog(
-            onDismiss = { showAddDialog = false },
-            onConfirm = { name, equipment, usesBodyweight, muscleGroups ->
-                viewModel.insert(name, equipment, usesBodyweight, muscleGroups)
-                showAddDialog = false
-            }
-        )
-    }
-
-    // Show the Edit dialog when exerciseToEdit is not null
-    if (exerciseToEdit != null) {
-        EditExerciseDialog(
-            exercise = exerciseToEdit!!,
-            onDismiss = { exerciseToEdit = null },
-            onConfirm = { updatedExercise ->
-                viewModel.update(updatedExercise)
-                exerciseToEdit = null
-            }
-        )
-    }
-
-    // Show the Delete confirmation dialog when exerciseToDelete is not null
-    if (exerciseToDelete != null) {
-        AlertDialog(
-            onDismissRequest = { exerciseToDelete = null },
-            title = { Text("Delete Exercise") },
-            text = { Text("Are you sure you want to delete '${exerciseToDelete!!.name}'? This cannot be undone.") },
-            confirmButton = {
-                Button(
-                    onClick = {
-                        viewModel.delete(exerciseToDelete!!)
-                        exerciseToDelete = null
-                    },
-                    colors = ButtonDefaults.buttonColors(containerColor = MaterialTheme.colorScheme.error)
-                ) { Text("Delete") }
-            },
-            dismissButton = {
-                TextButton(onClick = { exerciseToDelete = null }) { Text("Cancel") }
-            }
-        )
-    }
-}
-
-
-
-@Composable
-fun AddExerciseDialog(
-    onDismiss: () -> Unit,
-    onConfirm: (String, String, Boolean, List<MuscleGroup>) -> Unit
-) {
-    var newExerciseName by remember { mutableStateOf("") }
-    var newExerciseEquipment by remember { mutableStateOf("") }
-    var newExerciseUsesBodyweight by remember { mutableStateOf(false) }
-    var selectedMuscleGroups by remember { mutableStateOf<List<MuscleGroup>>(emptyList()) }
-    var showMuscleGroupDialog by remember { mutableStateOf(false) }
-
-    if (showMuscleGroupDialog) {
-        MuscleGroupSelectionDialog(
-            allMuscleGroups = MuscleGroup.entries.toList(),
-            initialSelection = selectedMuscleGroups,
-            onDismiss = { showMuscleGroupDialog = false },
-            onConfirm = {
-                selectedMuscleGroups = it
-                showMuscleGroupDialog = false
-            }
-        )
-    }
-
-    AlertDialog(
-        onDismissRequest = onDismiss,
-        title = { Text("Add New Exercise") },
-        text = {
-            Column(verticalArrangement = Arrangement.spacedBy(8.dp)) {
-                OutlinedTextField(
-                    value = newExerciseName,
-                    onValueChange = { newExerciseName = it },
-                    label = { Text("Exercise Name") }
-                )
-                OutlinedTextField(
-                    value = newExerciseEquipment,
-                    onValueChange = { newExerciseEquipment = it },
-                    label = { Text("Equipment (e.g., DUMBBELL)") }
-                )
-                Row(
-                    verticalAlignment = Alignment.CenterVertically,
-                    horizontalArrangement = Arrangement.SpaceBetween,
-                    modifier = Modifier.fillMaxWidth()
-                ) {
-                    Text("Uses Bodyweight")
-                    Switch(checked = newExerciseUsesBodyweight, onCheckedChange = { newExerciseUsesBodyweight = it })
-                }
-                Text("Target Muscles: ${selectedMuscleGroups.joinToString { it.name }}")
-                Button(onClick = { showMuscleGroupDialog = true }, modifier = Modifier.fillMaxWidth()) {
-                    Text("Select Muscle Groups")
-                }
-            }
-        },
-        confirmButton = {
-            Button(onClick = {
-                if (newExerciseName.isNotBlank()) {
-                    onConfirm(newExerciseName, newExerciseEquipment, newExerciseUsesBodyweight, selectedMuscleGroups)
-                }
-            }) { Text("Create") }
-        },
-        dismissButton = {
-            TextButton(onClick = onDismiss) { Text("Cancel") }
-        }
-    )
-}
-
-
-@Composable
-fun EditExerciseDialog(
-    exercise: Exercise,
-    onDismiss: () -> Unit,
-    onConfirm: (Exercise) -> Unit
-) {
-    var editedName by remember { mutableStateOf(exercise.name) }
-    var editedEquipment by remember { mutableStateOf(exercise.equipment.joinToString()) }
-    var editedUsesBodyweight by remember { mutableStateOf(exercise.usesBodyweight) }
-    var editedMuscleGroups by remember { mutableStateOf(exercise.targetMuscleGroups) }
-    var showMuscleGroupDialog by remember { mutableStateOf(false) }
-
-    if (showMuscleGroupDialog) {
-        MuscleGroupSelectionDialog(
-            allMuscleGroups = MuscleGroup.entries.toList(),
-            initialSelection = editedMuscleGroups,
-            onDismiss = { showMuscleGroupDialog = false },
-            onConfirm = {
-                editedMuscleGroups = it
-                showMuscleGroupDialog = false
-            }
-        )
-    }
-
-    AlertDialog(
-        onDismissRequest = onDismiss,
-        title = { Text("Edit Exercise") },
-        text = {
-            Column(verticalArrangement = Arrangement.spacedBy(8.dp)) {
-                OutlinedTextField(
-                    value = editedName,
-                    onValueChange = { editedName = it },
-                    label = { Text("Exercise Name") }
-                )
-                OutlinedTextField(
-                    value = editedEquipment,
-                    onValueChange = { editedEquipment = it },
-                    label = { Text("Equipment (e.g., DUMBBELL)") }
-                )
-                Row(
-                    verticalAlignment = Alignment.CenterVertically,
-                    horizontalArrangement = Arrangement.SpaceBetween,
-                    modifier = Modifier.fillMaxWidth()
-                ) {
-                    Text("Uses Bodyweight")
-                    Switch(checked = editedUsesBodyweight, onCheckedChange = { editedUsesBodyweight = it })
-                }
-                Text("Target Muscles: ${editedMuscleGroups.joinToString { it.name }}")
-                Button(onClick = { showMuscleGroupDialog = true }, modifier = Modifier.fillMaxWidth()) {
-                    Text("Select Muscle Groups")
-                }
-            }
-        },
-        confirmButton = {
-            Button(onClick = {
-                val updatedExercise = exercise.copy(
-                    name = editedName,
-                    equipment = try { listOf(Equipment.valueOf(editedEquipment.uppercase().trim())) } catch (e: Exception) { exercise.equipment },
-                    usesBodyweight = editedUsesBodyweight,
-                    targetMuscleGroups = editedMuscleGroups
-                )
-                onConfirm(updatedExercise)
-            }) { Text("Save") }
-        },
-        dismissButton = {
-            TextButton(onClick = onDismiss) { Text("Cancel") }
-        }
-    )
-}
-
-
-// In MainActivity.kt, add this new Composable function
-
-@Composable
-fun MuscleGroupSelectionDialog(
-    allMuscleGroups: List<MuscleGroup>,
-    initialSelection: List<MuscleGroup>,
-    onDismiss: () -> Unit,
-    onConfirm: (List<MuscleGroup>) -> Unit
-) {
-    var selectedMuscleGroups by remember { mutableStateOf(initialSelection.toSet()) }
-
-    AlertDialog(
-        onDismissRequest = onDismiss,
-        title = { Text("Select Muscle Groups") },
-        text = {
-            LazyColumn {
-                items(allMuscleGroups) { muscleGroup ->
-                    Row(
-                        verticalAlignment = Alignment.CenterVertically,
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .clickable {
-                                val newSelection = selectedMuscleGroups.toMutableSet()
-                                if (selectedMuscleGroups.contains(muscleGroup)) {
-                                    newSelection.remove(muscleGroup)
-                                } else {
-                                    newSelection.add(muscleGroup)
-                                }
-                                selectedMuscleGroups = newSelection
-                            }
-                            .padding(vertical = 4.dp)
-                    ) {
-                        Checkbox(
-                            checked = selectedMuscleGroups.contains(muscleGroup),
-                            onCheckedChange = { isChecked ->
-                                val newSelection = selectedMuscleGroups.toMutableSet()
-                                if (isChecked) {
-                                    newSelection.add(muscleGroup)
-                                } else {
-                                    newSelection.remove(muscleGroup)
-                                }
-                                selectedMuscleGroups = newSelection
-                            }
-                        )
-                        Spacer(modifier = Modifier.width(8.dp))
-                        Text(muscleGroup.name.replace("_", " ").lowercase()
-                            .replaceFirstChar { it.titlecase() })
-                    }
-                }
-            }
-        },
-        confirmButton = {
-            Button(onClick = { onConfirm(selectedMuscleGroups.toList()) }) {
-                Text("Confirm")
-            }
-        },
-        dismissButton = {
-            TextButton(onClick = onDismiss) { Text("Cancel") }
-        }
-    )
-}
-
-
-@OptIn(ExperimentalLayoutApi::class) // Needed for FlowRow
-@Composable
-fun ExerciseItem(
-    exercise: Exercise,
-    onEditClick: () -> Unit,
-    onDeleteClick: () -> Unit
-) {
-    Card(
-        modifier = Modifier.fillMaxWidth(),
-        elevation = CardDefaults.cardElevation(2.dp)
-    ) {
-        Column(modifier = Modifier.padding(16.dp)) {
-            Row(
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                // Exercise Name
-                Text(
-                    text = exercise.name,
-                    fontSize = 18.sp,
-                    fontWeight = FontWeight.Bold,
-                    modifier = Modifier.weight(1f)
-                )
-                // Action Buttons
-                IconButton(onClick = onEditClick) {
-                    Icon(Icons.Filled.Edit, contentDescription = "Edit Exercise")
-                }
-                IconButton(onClick = onDeleteClick) {
-                    Icon(Icons.Filled.Delete, contentDescription = "Delete Exercise")
-                }
-            }
-            Spacer(modifier = Modifier.height(8.dp))
-
-            // NEW: Display Muscle Groups as wrapping tags
-            FlowRow(
-                horizontalArrangement = Arrangement.spacedBy(4.dp),
-                verticalArrangement = Arrangement.spacedBy(4.dp)
-            ) {
-                exercise.targetMuscleGroups.forEach { group ->
-                    Text(
-                        text = group.name.replace("_", " ").lowercase()
-                            .replaceFirstChar { it.titlecase() },
-                        style = MaterialTheme.typography.bodySmall,
-                        modifier = Modifier
-                            .clip(RoundedCornerShape(4.dp))
-                            .background(MaterialTheme.colorScheme.secondaryContainer)
-                            .padding(horizontal = 6.dp, vertical = 2.dp)
-                    )
-                }
-            }
-        }
-    }
-}
-
-
-
-
-@OptIn(ExperimentalLayoutApi::class, ExperimentalMaterial3Api::class)
-@Composable
-fun VolumeAnalysisScreen(viewModel: VolumeViewModel) {
-    val volumeData by viewModel.volumeData.collectAsStateWithLifecycle()
-    val weeksInCycle by viewModel.weeksInActiveCycle.collectAsStateWithLifecycle()
-    val selectedWeek by viewModel.selectedWeek.collectAsStateWithLifecycle()
-
-    var isDropdownExpanded by remember { mutableStateOf(false) }
-
-    // Automatically select the first week when the screen loads
-    LaunchedEffect(weeksInCycle) {
-        if (selectedWeek == null) {
-            viewModel.onWeekSelected(weeksInCycle.firstOrNull())
-        }
-    }
-
-    val sortedVolumeList = volumeData.entries.toList().sortedByDescending { it.value }
-
-    val chartModelProducer = remember { ChartEntryModelProducer() }
-    val chartScrollState = rememberChartScrollState()
-
-    val bottomAxisValueFormatter = AxisValueFormatter<AxisPosition.Horizontal.Bottom> { value, _ ->
-        sortedVolumeList.getOrNull(value.toInt())?.key?.name
-            ?.replace("_", " ")?.lowercase()
-            ?.replaceFirstChar { if (it.isLowerCase()) it.titlecase(Locale.ROOT) else it.toString() }
-            ?: ""
-    }
-
-    val startAxisValueFormatter = AxisValueFormatter<AxisPosition.Vertical.Start> { value, _ ->
-        // Only display a label if the value is a whole number
-        if (value % 1 == 0f) {
-            value.toInt().toString()
-        } else {
-            "" // Return an empty string for decimal values
-        }
-    }
-
-    LaunchedEffect(sortedVolumeList) {
-        val entries = sortedVolumeList.mapIndexed { index, entry ->
-            entryOf(index.toFloat(), entry.value)
-        }
-        chartModelProducer.setEntries(entries)
-    }
-
-    Column(modifier = Modifier.padding(16.dp)) {
-        Text("Volume Analysis", fontSize = 24.sp, fontWeight = FontWeight.Bold)
-        Spacer(modifier = Modifier.height(8.dp))
-
-        // NEW: Dropdown menu for selecting a training week
-        ExposedDropdownMenuBox(
-            expanded = isDropdownExpanded,
-            onExpandedChange = { isDropdownExpanded = !isDropdownExpanded }
-        ) {
-            OutlinedTextField(
-                value = selectedWeek?.weekLabel ?: "Select a Week",
-                onValueChange = {},
-                readOnly = true,
-                trailingIcon = { ExposedDropdownMenuDefaults.TrailingIcon(expanded = isDropdownExpanded) },
-                modifier = Modifier.menuAnchor().fillMaxWidth()
-            )
-            ExposedDropdownMenu(
-                expanded = isDropdownExpanded,
-                onDismissRequest = { isDropdownExpanded = false }
-            ) {
-                weeksInCycle.forEach { week ->
-                    DropdownMenuItem(
-                        text = { Text(week.weekLabel) },
-                        onClick = {
-                            viewModel.onWeekSelected(week)
-                            isDropdownExpanded = false
-                        }
-                    )
-                }
-            }
-        }
-        Spacer(modifier = Modifier.height(16.dp))
-
-        if (volumeData.isEmpty()) {
-            Text("No workout data for the selected week to analyze.")
-        } else {
-            Card(elevation = CardDefaults.cardElevation(2.dp)) {
-                Column(
-                    modifier = Modifier.padding(16.dp)
-                ) {
-                    Chart(
-                        chart = columnChart(
-                            columns = listOf(
-                                lineComponent(
-                                    color = MaterialTheme.colorScheme.primary,
-                                    thickness = 12.dp,
-                                    dynamicShader = verticalGradient(
-                                        arrayOf(MaterialTheme.colorScheme.primary, MaterialTheme.colorScheme.secondary),
-                                    )
-                                )
-                            ),
-                            spacing = 16.dp
-                        ),
-                        chartModelProducer = chartModelProducer,
-                        startAxis = rememberStartAxis(
-                            title = "Total Sets",
-                            valueFormatter = startAxisValueFormatter
-                        ),
-                        bottomAxis = rememberBottomAxis(
-                            valueFormatter = bottomAxisValueFormatter,
-                            guideline = null
-                        ),
-                        modifier = Modifier.height(300.dp),
-                        chartScrollState = chartScrollState,
-                        isZoomEnabled = false
-                    )
-                    HorizontalDivider(modifier = Modifier.padding(vertical = 16.dp))
-                    Text("Details", style = MaterialTheme.typography.titleMedium)
-                    Spacer(modifier = Modifier.height(8.dp))
-
-                    LazyColumn(modifier = Modifier.heightIn(max = 200.dp)) {
-                        items(sortedVolumeList) { (muscleGroup, setCount) ->
-                            Row(
-                                modifier = Modifier.fillMaxWidth(),
-                                horizontalArrangement = Arrangement.SpaceBetween
-                            ) {
-                                Text(
-                                    muscleGroup.name.replace("_", " ").lowercase()
-                                        .replaceFirstChar { it.titlecase() },
-                                    style = MaterialTheme.typography.bodyLarge
-                                )
-                                Text("$setCount sets", style = MaterialTheme.typography.bodyLarge, fontWeight = FontWeight.Bold)
-                            }
-                            Spacer(modifier = Modifier.height(4.dp))
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-
-
-// A helper function to format seconds into MM:SS format
-private fun formatTime(seconds: Int): String {
-    val minutes = seconds / 60
-    val remainingSeconds = seconds % 60
-    return "%02d:%02d".format(minutes, remainingSeconds)
-}
-
-// The new UI component for the timer bar at the bottom of the screen
-@Composable
-fun TimerBar(
-    isRunning: Boolean,
-    currentTime: Int,
-    onPause: () -> Unit,
-    onResume: () -> Unit,
-    onStop: () -> Unit,
-    onReset: () -> Unit,
-    onAddTime: () -> Unit
-) {
-    AnimatedVisibility(visible = isRunning || currentTime > 0) {
-        BottomAppBar(
-            containerColor = MaterialTheme.colorScheme.surfaceVariant
-        ) {
-            Row(
-                modifier = Modifier.fillMaxWidth(),
-                horizontalArrangement = Arrangement.SpaceAround,
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                Text(
-                    text = formatTime(currentTime),
-                    style = MaterialTheme.typography.headlineMedium,
-                    fontWeight = FontWeight.Bold
-                )
-
-                // NEW: Dynamic Play/Pause Button
-                IconButton(onClick = { if (isRunning) onPause() else onResume() }) {
-                    Icon(
-                        imageVector = if (isRunning) Icons.Filled.PauseCircle else Icons.Filled.PlayArrow,
-                        contentDescription = if (isRunning) "Pause Timer" else "Resume Timer",
-                        modifier = Modifier.size(36.dp)
-                    )
-                }
-
-                Button(onClick = onAddTime) { Text("+15s") }
-
-                IconButton(onClick = onReset) {
-                    Icon(Icons.Filled.Replay, contentDescription = "Reset Timer")
-                }
-
-                IconButton(onClick = onStop) {
-                    Icon(Icons.Filled.Stop, contentDescription = "Stop Timer")
-                }
-            }
         }
     }
 }
