+++ app/src/main/java/com/example/myworkoutlog/ProgramManagementScreens.kt
@@ -1,0 +1,295 @@
+@file:OptIn(ExperimentalMaterial3Api::class)
+
+package com.example.myworkoutlog
+
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.items
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.*
+import androidx.compose.material.icons.automirrored.filled.ArrowBack
+import androidx.compose.material.icons.outlined.PlayArrow
+import androidx.compose.material3.*
+import androidx.compose.runtime.*
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import java.util.*
+
+@Composable
+fun ManageProgramsScreen(
+    programViewModel: ProgramViewModel,
+    activeCycleViewModel: ActiveCycleViewModel,
+    onNavigateToProgram: (String) -> Unit
+) {
+    val programs by programViewModel.allPrograms.collectAsStateWithLifecycle()
+    var showCreateProgramDialog by remember { mutableStateOf(false) }
+    var showStartCycleDialog by remember { mutableStateOf<ProgramTemplate?>(null) }
+    var newName by remember { mutableStateOf("") }
+
+    Scaffold(
+        floatingActionButton = {
+            FloatingActionButton(onClick = { showCreateProgramDialog = true }) {
+                Icon(Icons.Filled.Add, contentDescription = "Create new program")
+            }
+        }
+    ) { paddingValues ->
+        Column(modifier = Modifier.padding(paddingValues).padding(16.dp)) {
+            Text("Program Blueprints", fontSize = 24.sp, fontWeight = FontWeight.Bold)
+            Spacer(modifier = Modifier.height(16.dp))
+
+            if (programs.isEmpty()) {
+                Text(
+                    "No programs yet. Click the '+' button to create one.",
+                    modifier = Modifier.align(Alignment.CenterHorizontally),
+                    textAlign = TextAlign.Center
+                )
+            } else {
+                LazyColumn(verticalArrangement = Arrangement.spacedBy(8.dp)) {
+                    items(programs) { program ->
+                        Card(elevation = CardDefaults.cardElevation(2.dp)) {
+                            Row(
+                                modifier = Modifier.padding(start = 16.dp),
+                                verticalAlignment = Alignment.CenterVertically
+                            ) {
+                                Column(
+                                    modifier = Modifier
+                                        .weight(1f)
+                                        .clickable { onNavigateToProgram(program.id) }
+                                        .padding(vertical = 16.dp)
+                                ) { Text(program.name) }
+                                // "Start Cycle" button
+                                IconButton(onClick = { showStartCycleDialog = program }) {
+                                    Icon(
+                                        Icons.Outlined.PlayArrow,
+                                        contentDescription = "Start Cycle"
+                                    )
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+
+            // Dialog for creating a new program
+            if (showCreateProgramDialog) {
+                AlertDialog(
+                    onDismissRequest = { showCreateProgramDialog = false },
+                    title = { Text("New Program Blueprint") },
+                    text = {
+                        OutlinedTextField(
+                            value = newName,
+                            onValueChange = { newName = it },
+                            label = { Text("Program Name") },
+                            singleLine = true
+                        )
+                    },
+                    confirmButton = {
+                        Button(
+                            onClick = {
+                                if (newName.isNotBlank()) {
+                                    programViewModel.insert(newName, null)
+                                    newName = ""
+                                    showCreateProgramDialog = false
+                                }
+                            }
+                        ) { Text("Create") }
+                    },
+                    dismissButton = {
+                        TextButton(onClick = { showCreateProgramDialog = false }) {
+                            Text("Cancel")
+                        }
+                    }
+                )
+            }
+
+            // Dialog for starting a new cycle
+            if (showStartCycleDialog != null) {
+                val programToStart = showStartCycleDialog!!
+                AlertDialog(
+                    onDismissRequest = { showStartCycleDialog = null },
+                    title = { Text("Start New Cycle") },
+                    text = {
+                        Column {
+                            Text("You are about to start the program: ${programToStart.name}")
+                            Spacer(Modifier.height(8.dp))
+                            OutlinedTextField(
+                                value = newName,
+                                onValueChange = { newName = it },
+                                label = { Text("Give this cycle a name") },
+                                placeholder = { Text(programToStart.name) }
+                            )
+                        }
+                    },
+                    confirmButton = {
+                        Button(onClick = {
+                            val cycleName = newName.ifBlank { programToStart.name }
+                            activeCycleViewModel.startCycle(programToStart, cycleName)
+                            showStartCycleDialog = null
+                            newName = ""
+                        }) { Text("Start") }
+                    },
+                    dismissButton = {
+                        TextButton(onClick = {
+                            showStartCycleDialog = null
+                        }) { Text("Cancel") }
+                    }
+                )
+            }
+        }
+    }
+}
+
+@Composable
+fun ProgramEditorScreen(
+    programId: String,
+    programViewModel: ProgramViewModel,
+    templateViewModel: WorkoutTemplateViewModel,
+    onNavigateUp: () -> Unit
+) {
+    val programFromDb by programViewModel.getProgramById(programId).collectAsState(initial = null)
+    val allWorkoutTemplates by templateViewModel.allTemplates.collectAsStateWithLifecycle()
+
+    var editedName by remember { mutableStateOf("") }
+    var editedWeeks by remember { mutableStateOf<List<ProgramWeekDefinition>>(emptyList()) }
+    var showAddSessionDialog by remember { mutableStateOf<String?>(null) }
+
+    LaunchedEffect(programFromDb) {
+        programFromDb?.let {
+            editedName = it.name
+            editedWeeks = it.weeks
+        }
+    }
+
+    Scaffold(
+        topBar = {
+            TopAppBar(
+                title = { Text(editedName) },
+                navigationIcon = {
+                    IconButton(onClick = onNavigateUp) {
+                        Icon(Icons.AutoMirrored.Filled.ArrowBack, contentDescription = "Back")
+                    }
+                },
+                actions = {
+                    Button(onClick = {
+                        programFromDb?.let {
+                            val updatedProgram = it.copy(name = editedName, weeks = editedWeeks)
+                            programViewModel.update(updatedProgram)
+                        }
+                        onNavigateUp()
+                    }) { Text("Save") }
+                }
+            )
+        },
+        floatingActionButton = {
+            FloatingActionButton(onClick = {
+                val newWeek = ProgramWeekDefinition(
+                    id = UUID.randomUUID().toString(),
+                    weekLabel = "Week ${editedWeeks.size + 1}",
+                    order = editedWeeks.size + 1,
+                    sessions = emptyList()
+                )
+                editedWeeks = editedWeeks + newWeek
+            }) {
+                Icon(Icons.Filled.Add, "Add Week")
+            }
+        }
+    ) { paddingValues ->
+        Column(modifier = Modifier.padding(paddingValues).padding(horizontal = 16.dp)) {
+            OutlinedTextField(
+                value = editedName,
+                onValueChange = { editedName = it },
+                label = { Text("Program Name") },
+                modifier = Modifier.fillMaxWidth().padding(vertical = 8.dp)
+            )
+            LazyColumn(verticalArrangement = Arrangement.spacedBy(16.dp)) {
+                items(editedWeeks) { week ->
+                    Card(elevation = CardDefaults.cardElevation(2.dp)) {
+                        Column(Modifier.padding(16.dp)) {
+                            Row(verticalAlignment = Alignment.CenterVertically) {
+                                OutlinedTextField(
+                                    value = week.weekLabel,
+                                    onValueChange = { newLabel ->
+                                        editedWeeks = editedWeeks.map {
+                                            if (it.id == week.id) it.copy(weekLabel = newLabel) else it
+                                        }
+                                    },
+                                    label = { Text("Week Label") },
+                                    modifier = Modifier.weight(1f)
+                                )
+                                IconButton(onClick = {
+                                    editedWeeks = editedWeeks.filter { it.id != week.id }
+                                }) {
+                                    Icon(Icons.Default.Delete, contentDescription = "Delete Week")
+                                }
+                            }
+                            Spacer(Modifier.height(8.dp))
+                            // List the sessions for this week
+                            week.sessions.sortedBy { it.order }.forEach { session ->
+                                val templateName = allWorkoutTemplates.find { it.id == session.workoutTemplateId }?.name ?: "Unknown Template"
+                                Text("  - ${session.sessionName}: $templateName", style = MaterialTheme.typography.bodyMedium)
+                            }
+                            TextButton(onClick = { showAddSessionDialog = week.id }) {
+                                Text("Add Session to Week")
+                            }
+                        }
+                    }
+                }
+            }
+        }
+
+        // Dialog for adding a session to a week
+        if (showAddSessionDialog != null) {
+            val weekIdToAddSessionTo = showAddSessionDialog
+            var sessionName by remember { mutableStateOf("") }
+            AlertDialog(
+                onDismissRequest = { showAddSessionDialog = null },
+                title = { Text("Add Session") },
+                text = {
+                    Column {
+                        OutlinedTextField(
+                            value = sessionName,
+                            onValueChange = { sessionName = it },
+                            label = { Text("Session Name (e.g., Day 1)") }
+                        )
+                        Spacer(Modifier.height(8.dp))
+                        Text("Choose a template:", fontWeight = FontWeight.Bold)
+                        LazyColumn(modifier = Modifier.height(150.dp)) {
+                            items(allWorkoutTemplates) { template ->
+                                Text(
+                                    text = template.name,
+                                    modifier = Modifier
+                                        .fillMaxWidth()
+                                        .clickable {
+                                            val weekToUpdate = editedWeeks.find { it.id == weekIdToAddSessionTo }
+                                            if (weekToUpdate != null) {
+                                                val newSession = ProgramSessionDefinition(
+                                                    id = UUID.randomUUID().toString(),
+                                                    sessionName = sessionName.ifBlank { "Session ${weekToUpdate.sessions.size + 1}" },
+                                                    workoutTemplateId = template.id,
+                                                    order = weekToUpdate.sessions.size + 1
+                                                )
+                                                val updatedSessions = weekToUpdate.sessions + newSession
+                                                val updatedWeek = weekToUpdate.copy(sessions = updatedSessions)
+                                                editedWeeks = editedWeeks.map { if (it.id == weekIdToAddSessionTo) updatedWeek else it }
+                                            }
+                                            showAddSessionDialog = null
+                                        }
+                                        .padding(vertical = 8.dp)
+                                )
+                            }
+                        }
+                    }
+                },
+                confirmButton = {
+                    TextButton(onClick = { showAddSessionDialog = null }) { Text("Cancel") }
+                }
+            )
+        }
+    }
+}
