+++ app/src/main/java/com/example/myworkoutlog/SettingsScreen.kt
@@ -1,0 +1,41 @@
+package com.example.myworkoutlog
+
+import androidx.compose.foundation.layout.*
+import androidx.compose.material3.*
+import androidx.compose.runtime.*
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import androidx.lifecycle.compose.collectAsStateWithLifecycle
+
+@Composable
+fun SettingsScreen(viewModel: SettingsViewModel) {
+    val weightUnit by viewModel.weightUnit.collectAsStateWithLifecycle()
+    val options = listOf("kg", "lb")
+
+    Column(modifier = Modifier.padding(16.dp)) {
+        Text("Settings", fontSize = 24.sp, fontWeight = FontWeight.Bold)
+        Spacer(modifier = Modifier.height(16.dp))
+
+        Row(
+            modifier = Modifier.fillMaxWidth(),
+            verticalAlignment = Alignment.CenterVertically,
+            horizontalArrangement = Arrangement.SpaceBetween
+        ) {
+            Text("Weight Unit", style = MaterialTheme.typography.bodyLarge)
+            SingleChoiceSegmentedButtonRow {
+                options.forEachIndexed { index, label ->
+                    SegmentedButton(
+                        shape = SegmentedButtonDefaults.itemShape(index = index, count = options.size),
+                        onClick = { viewModel.setWeightUnit(label) },
+                        selected = weightUnit == label
+                    ) {
+                        Text(label)
+                    }
+                }
+            }
+        }
+    }
+}
