+++ app/src/main/java/com/example/myworkoutlog/WorkoutLoggerScreens.kt
@@ -1,0 +1,338 @@
+@file:OptIn(ExperimentalMaterial3Api::class)
+
+package com.example.myworkoutlog
+
+import androidx.compose.animation.AnimatedVisibility
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.items
+import androidx.compose.foundation.text.KeyboardOptions
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.*
+import androidx.compose.material.icons.automirrored.filled.ArrowBack
+import androidx.compose.material3.*
+import androidx.compose.runtime.*
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.focus.onFocusChanged
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.input.KeyboardType
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import androidx.lifecycle.compose.collectAsStateWithLifecycle
+
+@Composable
+fun WorkoutLoggerScreen(
+    templateId: String,
+    cycleId: String?,
+    weekId: String?,
+    sessionId: String?,
+    viewModel: WorkoutLoggerViewModel,
+    activeCycleViewModel: ActiveCycleViewModel,
+    weightUnit: String,
+    onNavigateUp: () -> Unit
+) {
+    // Local state for the bodyweight text field
+    var bodyweightText by remember { mutableStateOf("") }
+
+    // LaunchedEffect runs a coroutine when the composable first appears.
+    LaunchedEffect(key1 = templateId) {
+        viewModel.startWorkoutFromTemplate(templateId, cycleId, weekId, sessionId)
+    }
+
+    // Collect the active workout state from the ViewModel.
+    val activeWorkout by viewModel.activeWorkoutState.collectAsStateWithLifecycle()
+    val activeCycle by activeCycleViewModel.activeCycle.collectAsStateWithLifecycle()
+
+    // Get the timer state from the ViewModel
+    val timerIsRunning by viewModel.timerIsRunning.collectAsStateWithLifecycle()
+    val timerValue by viewModel.timerValueSeconds.collectAsStateWithLifecycle()
+    val sessionElapsedTime by viewModel.sessionElapsedTime.collectAsStateWithLifecycle()
+
+    Scaffold(
+        topBar = {
+            TopAppBar(
+                title = { Text(activeWorkout?.name ?: "Log Workout") },
+                navigationIcon = {
+                    IconButton(onClick = {
+                        viewModel.stopRestTimer()
+                        onNavigateUp()
+                    }) {
+                        Icon(Icons.AutoMirrored.Filled.ArrowBack, contentDescription = "Back")
+                    }
+                },
+                actions = {
+                    Text(
+                        text = formatTime(sessionElapsedTime),
+                        style = MaterialTheme.typography.titleMedium,
+                        modifier = Modifier.padding(end = 8.dp)
+                    )
+                    Button(onClick = {
+                        viewModel.finishWorkout(weightUnit, activeCycle)
+                        onNavigateUp()
+                    }) {
+                        Text("Finish")
+                    }
+                }
+            )
+        },
+
+        bottomBar = {
+            TimerBar(
+                isRunning = timerIsRunning,
+                currentTime = timerValue,
+                onPause = { viewModel.pauseRestTimer() },
+                onResume = { viewModel.resumeRestTimer() },
+                onStop = { viewModel.stopRestTimer() },
+                onReset = { viewModel.resetRestTimer() },
+                onAddTime = { viewModel.addTimeToRestTimer(15) }
+            )
+        }
+    ) { paddingValues ->
+        // If the workout is not loaded yet, show a loading indicator.
+        if (activeWorkout == null) {
+            Box(modifier = Modifier.fillMaxSize(), contentAlignment = Alignment.Center) {
+                CircularProgressIndicator()
+            }
+        } else {
+            // Once loaded, display the list of exercises.
+            LazyColumn(
+                modifier = Modifier
+                    .padding(paddingValues)
+                    .padding(16.dp),
+                verticalArrangement = Arrangement.spacedBy(16.dp)
+            ) {
+                // Add a bodyweight input field as the first item in the list
+                item {
+                    OutlinedTextField(
+                        value = bodyweightText,
+                        onValueChange = { newText ->
+                            if (newText.matches(Regex("^\\d*\\.?\\d*\$"))) {
+                                bodyweightText = newText
+                            }
+                        },
+                        label = { Text("Bodyweight ($weightUnit)") },
+                        modifier = Modifier
+                            .fillMaxWidth()
+                            .onFocusChanged { focusState ->
+                                if (!focusState.isFocused) {
+                                    viewModel.updateBodyweight(bodyweightText)
+                                }
+                            }
+                    )
+                }
+                items(activeWorkout!!.loggedExercises) { exercise ->
+                    Card(modifier = Modifier.fillMaxWidth()) {
+                        Column(modifier = Modifier.padding(16.dp)) {
+                            Text(exercise.exerciseName, fontSize = 18.sp, fontWeight = FontWeight.Bold)
+                            Spacer(Modifier.height(8.dp))
+                            HorizontalDivider(modifier = Modifier.padding(vertical = 8.dp))
+
+                            // For each exercise, display its sets
+                            exercise.sets.forEachIndexed { index, set ->
+                                LoggedSetRow(
+                                    set = set,
+                                    setNumber = index + 1,
+                                    weightUnit = weightUnit,
+                                    onRepsChange = { newReps ->
+                                        viewModel.updateSet(exercise.id, set.id, newReps, set.weight, set.secs?.toString() ?: "")
+                                    },
+                                    onWeightChange = { newWeight ->
+                                        viewModel.updateSet(exercise.id, set.id, set.reps?.toString() ?: "", newWeight, set.secs?.toString() ?: "")
+                                    },
+                                    onSecsChange = { newSecs ->
+                                        viewModel.updateSet(exercise.id, set.id, set.reps?.toString() ?: "", set.weight, newSecs)
+                                    },
+                                    // Pass the ViewModel function to the button
+                                    onStartRest = { viewModel.startRestTimer() }
+                                )
+                            }
+                        }
+                    }
+                }
+            }
+        }
+    }
+}
+
+@Composable
+fun LoggedSetRow(
+    set: LoggedSet,
+    setNumber: Int,
+    weightUnit: String,
+    onRepsChange: (String) -> Unit,
+    onWeightChange: (Double?) -> Unit,
+    onSecsChange: (String) -> Unit,
+    onStartRest: () -> Unit
+) {
+    // Determine which fields to show based on the template's targets for this set
+    val showWeightReps = !set.targetReps.isNullOrBlank()
+    val showSecs = !set.targetSecs.isNullOrBlank()
+
+    // Local state holds the text as the user types it.
+    var weightText by remember { mutableStateOf(set.weight?.toString() ?: "") }
+    var repsText by remember { mutableStateOf(set.reps?.toString() ?: "") }
+    var secsText by remember { mutableStateOf(set.secs?.toString() ?: "") }
+
+    // This ensures if the underlying data changes from elsewhere, our text fields update.
+    LaunchedEffect(set.weight) {
+        val currentWeightString = set.weight?.toString() ?: ""
+        if (weightText != currentWeightString) {
+            weightText = currentWeightString
+        }
+    }
+    LaunchedEffect(set.reps) {
+        val currentRepsString = set.reps?.toString() ?: ""
+        if (repsText != currentRepsString) {
+            repsText = currentRepsString
+        }
+    }
+    LaunchedEffect(set.secs) {
+        val currentSecsString = set.secs?.toString() ?: ""
+        if (secsText != currentSecsString) {
+            secsText = currentSecsString
+        }
+    }
+
+    Column(modifier = Modifier.padding(vertical = 4.dp)) {
+        Row(
+            modifier = Modifier.fillMaxWidth(),
+            verticalAlignment = Alignment.CenterVertically
+        ) {
+            Text(
+                text = "Set $setNumber",
+                modifier = Modifier.width(60.dp),
+                fontWeight = FontWeight.Bold
+            )
+            Spacer(modifier = Modifier.weight(1f))
+            if (set.reps != null || set.secs != null) {
+                IconButton(onClick = onStartRest) {
+                    Icon(Icons.Filled.PlayArrow, contentDescription = "Start Rest Timer")
+                }
+            }
+        }
+
+        Row(
+            modifier = Modifier.fillMaxWidth(),
+            verticalAlignment = Alignment.CenterVertically,
+            horizontalArrangement = Arrangement.spacedBy(8.dp)
+        ) {
+            // Text field for weight
+            OutlinedTextField(
+                value = weightText,
+                onValueChange = { newText ->
+                    // Allow only digits and one decimal point
+                    if (newText.matches(Regex("^\\d*\\.?\\d*\$"))) {
+                        weightText = newText
+                    }
+                },
+                label = { Text("Weight ($weightUnit)") },
+                keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Number),
+                modifier = Modifier
+                    .weight(1f)
+                    .onFocusChanged { focusState ->
+                        if (!focusState.isFocused) {
+                            onWeightChange(weightText.toDoubleOrNull())
+                        }
+                    }
+            )
+
+            if (showWeightReps) {
+                // Text field for reps
+                OutlinedTextField(
+                    value = repsText,
+                    onValueChange = { newText ->
+                        if (newText.all { it.isDigit() }) {
+                            repsText = newText
+                        }
+                    },
+                    label = { Text("Reps") },
+                    keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Number),
+                    modifier = Modifier
+                        .weight(1f)
+                        .onFocusChanged { focusState ->
+                            if (!focusState.isFocused) {
+                                onRepsChange(repsText)
+                            }
+                        }
+                )
+            }
+
+            // This TextField will only appear for time-based exercises
+            if (showSecs) {
+                OutlinedTextField(
+                    value = secsText,
+                    onValueChange = { newText ->
+                        if (newText.all { it.isDigit() }) {
+                            secsText = newText
+                        }
+                    },
+                    label = { Text("Secs") },
+                    keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Number),
+                    modifier = Modifier
+                        .weight(1f)
+                        .onFocusChanged { focusState ->
+                            if (!focusState.isFocused) {
+                                onSecsChange(secsText)
+                            }
+                        }
+                )
+            }
+        }
+    }
+}
+
+@Composable
+fun TimerBar(
+    isRunning: Boolean,
+    currentTime: Int,
+    onPause: () -> Unit,
+    onResume: () -> Unit,
+    onStop: () -> Unit,
+    onReset: () -> Unit,
+    onAddTime: () -> Unit
+) {
+    AnimatedVisibility(visible = isRunning || currentTime > 0) {
+        BottomAppBar(
+            containerColor = MaterialTheme.colorScheme.surfaceVariant
+        ) {
+            Row(
+                modifier = Modifier.fillMaxWidth(),
+                horizontalArrangement = Arrangement.SpaceAround,
+                verticalAlignment = Alignment.CenterVertically
+            ) {
+                Text(
+                    text = formatTime(currentTime),
+                    style = MaterialTheme.typography.headlineMedium,
+                    fontWeight = FontWeight.Bold
+                )
+
+                // Dynamic Play/Pause Button
+                IconButton(onClick = { if (isRunning) onPause() else onResume() }) {
+                    Icon(
+                        imageVector = if (isRunning) Icons.Filled.PauseCircle else Icons.Filled.PlayArrow,
+                        contentDescription = if (isRunning) "Pause Timer" else "Resume Timer",
+                        modifier = Modifier.size(36.dp)
+                    )
+                }
+
+                Button(onClick = onAddTime) { Text("+15s") }
+
+                IconButton(onClick = onReset) {
+                    Icon(Icons.Filled.Replay, contentDescription = "Reset Timer")
+                }
+
+                IconButton(onClick = onStop) {
+                    Icon(Icons.Filled.Stop, contentDescription = "Stop Timer")
+                }
+            }
+        }
+    }
+}
+
+// A helper function to format seconds into MM:SS format
+private fun formatTime(seconds: Int): String {
+    val minutes = seconds / 60
+    val remainingSeconds = seconds % 60
+    return "%02d:%02d".format(minutes, remainingSeconds)
+}
